Move shortTitle to base clash
