+++ src/all/hentaifox/src/eu/kanade/tachiyomi/extension/all/hentaifox/HentaiFox.kt
@@ -41,6 +41,8 @@ class HentaiFox(
             }
         }
 
+    override fun Element.mangaTitle(selector: String): String? = mangaFullTitle(selector)
+
     override fun Element.getTime(): Long {
         return selectFirst(".pages:contains(Posted:)")?.ownText()
             ?.removePrefix("Posted: ")
