+++ src/all/imhentai/src/eu/kanade/tachiyomi/extension/all/imhentai/IMHentai.kt
@@ -1,13 +1,8 @@
 package eu.kanade.tachiyomi.extension.all.imhentai
 
-import android.app.Application
-import android.content.SharedPreferences
-import androidx.preference.PreferenceScreen
-import androidx.preference.SwitchPreferenceCompat
 import eu.kanade.tachiyomi.multisrc.galleryadults.GalleryAdults
 import eu.kanade.tachiyomi.multisrc.galleryadults.cleanTag
 import eu.kanade.tachiyomi.multisrc.galleryadults.imgAttr
-import eu.kanade.tachiyomi.source.ConfigurableSource
 import okhttp3.FormBody
 import okhttp3.HttpUrl.Companion.toHttpUrl
 import okhttp3.OkHttpClient
@@ -15,14 +10,12 @@ import okhttp3.Response
 import okhttp3.ResponseBody.Companion.toResponseBody
 import org.jsoup.nodes.Document
 import org.jsoup.nodes.Element
-import uy.kohesive.injekt.Injekt
-import uy.kohesive.injekt.api.get
 import java.io.IOException
 
 class IMHentai(
     lang: String = "all",
     override val mangaLang: String = LANGUAGE_MULTI,
-) : ConfigurableSource, GalleryAdults(
+) : GalleryAdults(
     "IMHentai",
     "https://imhentai.xxx",
     lang = lang,
@@ -31,36 +24,6 @@ class IMHentai(
     override val useIntermediateSearch: Boolean = true
     override val supportAdvanceSearch: Boolean = true
     override val supportSpeechless: Boolean = true
-
-    private val preferences: SharedPreferences by lazy {
-        Injekt.get<Application>().getSharedPreferences("source_$id", 0x0000)
-    }
-
-    private val SharedPreferences.shortTitle
-        get() = getBoolean(PREF_SHORT_TITLE, false)
-
-    private val shortenTitleRegex = Regex("""(\[[^]]*]|[({][^)}]*[)}])""")
-
-    private fun String.shortenTitle() = this.replace(shortenTitleRegex, "").trim()
-
-    override fun setupPreferenceScreen(screen: PreferenceScreen) {
-        SwitchPreferenceCompat(screen.context).apply {
-            key = PREF_SHORT_TITLE
-            title = "Display Short Titles"
-            summaryOff = "Showing Long Titles"
-            summaryOn = "Showing short Titles"
-            setDefaultValue(false)
-        }.also(screen::addPreference)
-    }
-
-    override fun Element.mangaTitle(selector: String) =
-        mangaFullTitle(selector).let {
-            if (preferences.shortTitle) it?.shortenTitle() else it
-        }
-
-    private fun Element.mangaFullTitle(selector: String) =
-        selectFirst(selector)?.text()
-            ?.replace("\"", "")?.trim()
 
     override fun Element.mangaLang() =
         select("a:has(.thumb_flag)").attr("href")
@@ -199,8 +162,4 @@ class IMHentai(
     }
 
     override val idPrefixUri = "gallery"
-
-    companion object {
-        private const val PREF_SHORT_TITLE = "pref_short_title"
-    }
 }
